refactor: Reorganize directory structure.
