+++ src/helpers/desc.js
