+++ src/index.js
@@ -25,6 +25,13 @@ export { seed } from './util/random';
 export { default as fromArrow } from './format/from-arrow';
 export { default as fromCSV } from './format/from-csv';
 export { default as fromJSON } from './format/from-json';
-export { query, queryFrom } from './query/query';
+export { default as bin } from './helpers/bin';
+export { default as desc } from './helpers/desc';
+export { default as field } from './helpers/field';
+export { default as frac } from './helpers/frac';
+export { default as rolling } from './helpers/rolling';
+export { all, endswith, matches, not, range, startswith} from './helpers/selection';
+export { default as op } from './op/op-api';
 export * from './op/register';
-export * from './verbs';
+export { query, queryFrom } from './query/query';
+export * from './table';
