+++ src/verbs/fold.js
@@ -1,5 +1,5 @@
 import _fold from '../engine/fold';
-import parse from './expr/parse';
+import parse from './util/parse';
 
 export default function(table, values, options) {
   return _fold(table, parse('fold', table, values), options);
