+++ src/verbs/groupby.js
@@ -1,5 +1,5 @@
 import _groupby from '../engine/groupby';
-import parse from './expr/parse';
+import parse from './util/parse';
 
 export default function(table, values) {
   return _groupby(table, parse('groupby', table, values));
