+++ src/verbs/join.js
@@ -1,8 +1,8 @@
 import _join from '../engine/join';
-import parseKey from './expr/parse-key';
-import parseValue from './expr/parse';
-import { all, not } from './expr/selection';
+import parseKey from './util/parse-key';
+import parseValue from './util/parse';
 import parse from '../expression/parse';
+import { all, not } from '../helpers/selection';
 import error from '../util/error';
 import intersect from '../util/intersect';
 import isArray from '../util/is-array';
