+++ src/verbs/lookup.js
@@ -1,7 +1,7 @@
 import _lookup from '../engine/lookup';
 import { inferKeys } from './join';
-import parseKey from './expr/parse-key';
-import parseValues from './expr/parse';
+import parseKey from './util/parse-key';
+import parseValues from './util/parse';
 
 export default function(tableL, tableR, on, values) {
   on = inferKeys(tableL, tableR, on);
