+++ src/verbs/select.js
@@ -1,5 +1,5 @@
 import _select from '../engine/select';
-import resolve from './expr/selection';
+import resolve from '../helpers/selection';
 
 export default function(table, columns) {
   return _select(table, resolve(table, columns));
