+++ src/verbs/spread.js
@@ -1,5 +1,5 @@
 import _spread from '../engine/spread';
-import parse from './expr/parse';
+import parse from './util/parse';
 
 export default function(table, values, options) {
   return _spread(table, parse('spread', table, values), options);
