+++ src/verbs/util/parse.js
@@ -1,6 +1,6 @@
-import field from './field';
-import resolve from './selection';
 import parse from '../../expression/parse';
+import field from '../../helpers/field';
+import resolve from '../../helpers/selection';
 import assign from '../../util/assign';
 import error from '../../util/error';
 import isNumber from '../../util/is-number';
