+++ test/verbs/union-test.js
@@ -1,6 +1,6 @@
 import tape from 'tape';
 import tableEqual from '../table-equal';
-import { table } from '../../src/verbs';
+import { table } from '../../src';
 
 tape('union combines tables', t => {
   const t1 = table({ a: [1, 2], b: [3, 4] });
